+++ dal/validation/validator.py
@@ -1,105 +1,68 @@
 """
-   Copyright (C) Mov.ai  - All Rights Reserved
-   Unauthorized copying of this file, via any medium is strictly prohibited
-   Proprietary and confidential
+Copyright (C) Mov.ai  - All Rights Reserved
+Unauthorized copying of this file, via any medium is strictly prohibited
+Proprietary and confidential
 
-   Developers:
-   - Moawiya Mograbi (moawiya@mov.ai) - 2022
+Developers:
+- Moawiya Mograbi (moawiya@mov.ai) - 2022
 """
 
-from os import listdir
-from os.path import dirname, realpath, isdir
-from pathlib import Path
 from re import search
+import urllib
+from typing import Dict
+from pathlib import Path
 
 from .schema import Schema
-from json import loads as load_json
-from dal.exceptions import SchemaTypeNotKnown, ValidationError, SchemaVersionError
+from dal.exceptions import SchemaTypeNotKnown, SchemaVersionError
+from .constants import SCHEMA_FOLDER_PATH
+from dal.classes.common.singleton import Singleton
 
 
-class JsonValidator:
-    """JsonValidator class
-    responsible to load schema json files and validate files according
+class JsonValidator(metaclass=Singleton):
+    """Validator responsible to load schema json files and validate files according
     to it's type.
-    types: node/flow/callback/annotation/layout/graphicscene
+
     """
 
-    def __init__(self, version: str):
-        self.schema_types = []
-        self._version = version
+    VERSION = "2.4"
+
+    def __init__(self):
+        self.schema_types: Dict[str, Schema] = {}
         self._init_schemas()
 
-    def load_schema(self, schema_type: str, schema_path: str) -> bool:
-        """load single schema in Validator class, check versions if bigger
-           than existing one then replace it.
-
-        Args:
-            schema_type (str): schema type, node/flow/callback...
-            schema_path (str): path of the schema file to be loaded
-
-        Returns:
-            bool: True if loaded successfully, otherwise False.
-        """
-        schema_type = schema_type.lower()
-        schema_obj = Schema(schema_path)
-        if schema_type in self.schema_types:
-            # schema already loaded
-            curr_schema: Schema = getattr(self, schema_type)
-            if curr_schema.version > schema_obj.version:
-                # TODO print or raise approperiate message
-                return False
-        else:
-            self.schema_types.append(schema_type)
-        setattr(self, schema_type, schema_obj)
-
-        return True
-
     def _init_schemas(self):
-        """will initialize schemas objects in the schema folder
-        for all of our configuration files
-        """
-        # validator is in the root of validation module
-        validation_path = dirname(realpath(__file__))
+        """Initialize schemas objects in the schema folder for all of our configuration files."""
+        schema_folder = Path(
+            urllib.parse.urlparse(SCHEMA_FOLDER_PATH).path
+        )  # remove 'file://' prefix
+        version_folder = schema_folder / self.VERSION
 
-        schema_version_folder = f"{validation_path}/schema/{self._version}"
-        if not isdir(schema_version_folder):
-            raise SchemaVersionError(f"version {self._version} does not exist in schema folder")
-        for schema_json in listdir(schema_version_folder):
-            m = search(r"(\w+)\.schema\.json", schema_json)
+        if not version_folder.exists():
+            raise SchemaVersionError(f"Version folder {version_folder} does not exist")
+
+        for schema_json in version_folder.iterdir():
+            m = search(r"(\w+)\.schema\.json", schema_json.name)
             if m is not None:
-                schema_type = m.group(1).lower()
-                schema_path = f"{schema_version_folder}/{schema_json}"
-                if not self.load_schema(schema_type, schema_path):
-                    # loading failed
-                    # TODO print approperiate message
-                    pass
+                schema_type = m.group(1)
+                self.schema_types[schema_type] = Schema(schema_json)
 
-    def validate(self, file_path: Path, content=None) -> dict:
-        """validate a local file path against it's matching schema
+    def validate(self, scope: str, data: dict):
+        """Validate the content against the schema of the given scope.
 
         Args:
-            file_path (str): the local file path to be checked
+            scope (str): The type of the schema to validate against.
+            data (dict): The data to validate.
 
-        Returns:
-            dict: a dictionary including a status about the validation
-                  same as Schema.validate return value
-                        - status: True if succeeded otherwise False
-                        - message: error or success message
-                        - path: the path of the error in case there is one
-        """
-        if file_path is not None:
-            with file_path.open() as f:
-                content = load_json(f.read())
-        type = (list(content.keys())[0]).lower()
-        if type not in self.schema_types:
-            raise SchemaTypeNotKnown(f"type: {type}")
-        schema_obj: Schema = getattr(self, type)
+        Raises:
+            SchemaTypeNotKnown: If the scope is not known to the validator.
+            ValueError: If the data does not conform to the schema.
 
         """
-        validation_res = schema_obj.validate(content)
-        if validation_res["status"] is False:
-            # validation Failed
-            raise ValidationError(f"message:{validation_res['message']},\
-                                path:{validation_res['path']}")
-        """
-        return schema_obj.validate(content)
+        if scope not in self.schema_types:
+            raise SchemaTypeNotKnown(
+                f"Type {scope} is not known to the validator: {self.schema_types}"
+            )
+
+        result = self.schema_types[scope].validate(data)
+        if not result["status"]:
+            raise ValueError(f"Invalid data for scope {scope}: {result['message']} for data {data}")